Fixed bot message being a paragraph. Added bot parameters. Fixed toggle furniture not toggling triggering item. Rolling onto gates now works. Teleporting (teleport cmd) to furniture that can be sit on now updates the user

+++ src/main/java/com/eu/habbo/habbohotel/catalog/CatalogManager.java
@@ -784,9 +784,9 @@ public class CatalogManager
             level = 2;
         }
 
-        if (this.prizes.containsKey((Integer)level) && !this.prizes.get((Integer)level).isEmpty())
+        if (this.prizes.containsKey(level) && !this.prizes.get(level).isEmpty())
         {
-            return (Item) this.prizes.get((Integer) level).toArray()[Emulator.getRandom().nextInt(this.prizes.get(level).size())];
+            return (Item) this.prizes.get(level).toArray()[Emulator.getRandom().nextInt(this.prizes.get(level).size())];
         }
         else
         {
